+++ src/templates/omni.ts
@@ -15,14 +15,14 @@ export default (config: GenerateOmniConfigParams) => {
 
 const path = require('path');
 ${project_type === 'spa_react' ? `const merge = require('webpack-merge');
-const prod_config = require('./configs/webpack.config.prod.js');` : ''}
-${project_type !== 'component_library_react' ? 'const dev_config = require(\'./configs/webpack.config.dev.js\');\n' : ''}
+const config_prod = require('./configs/webpack.config.prod.js');` : ''}
+${project_type !== 'component_library_react' ? 'const config_dev = require(\'./configs/webpack.config.dev.js\');\n' : ''}
 module.exports = {
   type: '${project_type}', // 项目类型，请勿变动 (project type, please don't modify)
   ${project_type !== 'component_library_react' ? `\n  dev: {
-    webpack_config: dev_config, // 开发服务端webpack配置文件路径 (dev-server webpack config path)
+    webpack: config_dev, // 开发服务端webpack配置文件路径 (dev-server webpack config path)
     proxy:  [
-      // { 
+      // {
       //   route: '/api',
       //   config: {
       //     target: 'http://www.api.com/api',
@@ -30,7 +30,8 @@ module.exports = {
       //   }
       // }
     ], // 开发服务代理配置 (dev-server proxy config)
-    port: 6200 // 开发服务端口号 (dev-server port)
+    port: 6200, // 开发服务端口号 (dev-server port)
+    logLevel: 'error' // 开发服务日志输出等级，可选 'debug'、'info'、'warn'、'error'、'silent' (The log-level which dev-server will apply)
   },\n` : '' }
   build: {
     auto_release: false, // 构建完成后是否自动发布 (auto release project after build success)
@@ -45,7 +46,7 @@ module.exports = {
     esm_dir: path.resolve('es'),` : ''}
     ${project_type !== 'component_library_react' ? `// (构建阶段的自定义配置回调) The callback will be call in the build-process
     // (返回自定义的配置) You can return your custom build configuration
-    ${project_type === 'spa_react' ? 'configuration: config => merge(config, prod_config)' : 'configuration: config => config'},` : ''}
+    ${project_type === 'spa_react' ? 'configuration: config => merge(config, config_prod)' : 'configuration: config => config'},` : ''}
     ${project_type === 'toolkit' ? `tool: '${build}', // 打包工具，支持 tsc、rollup、webpack (build tool, support tsc, rollup and webpack)` : ''}
     reserve: {
       style: ${style && build !== 'webpack' ? true : false}, // 构建结果是否保留样式文件 (whether or not reserve the stylesheet files)